Github Repos/Signify_backend/requirements.txt
echo "fastapi
uvicorn
tensorflow
pillow
" > backend/requirements.txt
